+++ packages/react/lock-modular-temp/src/createFocusScope.tsx
