+++ packages/react/lock-modular-temp/src/index.ts
@@ -1,1 +1,1 @@
-export * from './FocusTrap';
+export * from './FocusScope';
